chunker noun defaults and bug fixes
