--- nlu/components/utils/ner_to_chunk_converter/__init__.py
@@ -1,11 +1,0 @@
-import nlu.pipe_components
-import sparknlp
-from sparknlp.base import *
-from sparknlp.annotator import *
-
-class NerToChunkConverter:
-    @staticmethod
-    def get_default_model():
-        return NerConverter() \
-            .setInputCols(["sentence", "token", "ner"]) \
-            .setOutputCol("ner_chunk")
