+++ nlu/components/utils/ner_to_chunk_converter/component_infos.json
@@ -6,8 +6,8 @@
   "type" : "util",
   "file_dependencies" : {},
   "pipe_priorioty_slot" : "2" ,
-  "spark_input_column_names" :["chunk_output" ] ,
-  "spark_output_column_names" :["sentence", "token", "ner"],
+  "spark_input_column_names" :["sentence", "token", "ner" ] ,
+  "spark_output_column_names" :["ner_chunk"],
   "critical_parameters" :[],
   "tags" : ["embeddings", "transformer", "deep learning"],
   "categories" : ["embeddings", "contextual embeddings"],
