+++ nlu/components/utils/ner_to_chunk_converter/ner_to_chunk_converter.py
@@ -1,0 +1,12 @@
+import nlu.pipe_components
+import sparknlp
+from sparknlp.base import *
+from sparknlp.annotator import *
+
+class NerToChunkConverter:
+    @staticmethod
+    def get_default_model():
+        return NerConverter() \
+            .setInputCols(["sentence", "token", "ner"]) \
+            .setOutputCol("ner_chunk")
+
